+++ src/CoreOptions.php
@@ -203,6 +203,41 @@ class CoreOptions {
                 'The maximum number of files that will be crawled at the same time.'
             ),
             self::makeOptionSpec(
+                'fileExtensionsToIgnore',
+                '1',
+                'File Extensions to Ignore',
+                'Files with these extensions will be ignored while crawling.',
+                implode(
+                    "\n",
+                    [
+                        '.bat',
+                        '.crt',
+                        '.DS_Store',
+                        '.git',
+                        '.idea',
+                        '.ini',
+                        '.less',
+                        '.map',
+                        '.md',
+                        '.mo',
+                        '.php',
+                        '.PHP',
+                        '.phtml',
+                        '.po',
+                        '.pot',
+                        '.scss',
+                        '.sh',
+                        '.sql',
+                        '.SQL',
+                        '.tar.gz',
+                        '.tpl',
+                        '.txt',
+                        '.yarn',
+                        '.zip',
+                    ]
+                )
+            ),
+            self::makeOptionSpec(
                 'filenamesToIgnore',
                 '1',
                 'Directory and File Names to Ignore',
@@ -662,6 +697,17 @@ VALUES (%s, %s, %s);";
                     $table_name,
                     [ 'value' => $crawl_concurrency < 1 ? 1 : $crawl_concurrency ],
                     [ 'name' => 'crawlConcurrency' ]
+                );
+
+                $file_extensions_to_ignore = preg_replace(
+                    '/^\s+|\s+$/m',
+                    '',
+                    $_POST['fileExtensionsToIgnore']
+                );
+                $wpdb->update(
+                    $table_name,
+                    [ 'blob_value' => $file_extensions_to_ignore ],
+                    [ 'name' => 'fileExtensionsToIgnore' ]
                 );
 
                 $filenames_to_ignore = preg_replace(