+++ views/advanced-options-page.php
@@ -36,6 +36,23 @@
                         ><?php echo $view['coreOptions']['filenamesToIgnore']->blob_value; ?></textarea>
                 </td>
             </tr>
+            <tr>
+                <td style="width:50%;">
+                    <label
+                        for="<?php echo $view['coreOptions']['fileExtensionsToIgnore']->name; ?>"
+                    ><b><?php echo $view['coreOptions']['fileExtensionsToIgnore']->label; ?></b></label>
+                    <br/><?php echo $view['coreOptions']['fileExtensionsToIgnore']->description; ?>
+                </td>
+                <td>
+                    <textarea
+                        class="widefat"
+                        cols=30 rows=10
+                        id="<?php echo $view['coreOptions']['fileExtensionsToIgnore']->name; ?>"
+                        name="<?php echo $view['coreOptions']['fileExtensionsToIgnore']->name; ?>"
+                        type="text"
+                        ><?php echo $view['coreOptions']['fileExtensionsToIgnore']->blob_value; ?></textarea>
+                </td>
+            </tr>
         </tbody>
     </table>